ref: trainer argparse 1/n (#3421)
* ref: trainer argparse 1/n

* ref: trainer argparse 1/n

* ref: trainer argparse 1/n

* ref: trainer argparse 1/n

* ref: trainer argparse 1/n

* ref: trainer argparse 1/n

* ref: trainer argparse 1/n

* ref: trainer argparse 1/n
